test: rename test apps
Having the Django and Flask apps both use the module name "app" makes
the tests order-dependent. If the Django "app" is loaded before the
Flask tests, the latter fail because they try to import "app.main",
which doesn't exist

Maybe there's some other way to avoid this collision by unloading "app"
and still have all the tests pass, but I couldn't find it. Just renaming
them seemed simplest.
