+++ _appmap/test/data/django/djangoapp/__init__.py
@@ -6,7 +6,7 @@ import django.conf
 django.conf.settings.configure(
     DATABASES={"default": {"ENGINE": "django.db.backends.sqlite3", "NAME": ":memory:"}},
     MIDDLEWARE=["django.middleware.http.ConditionalGetMiddleware"],
-    ROOT_URLCONF="app.urls",
+    ROOT_URLCONF="djangoapp.urls",
     TEMPLATES=[
         {
             "BACKEND": "django.template.backends.django.DjangoTemplates",
