+++ _appmap/test/data/django/djangoapp/settings.py
@@ -6,7 +6,7 @@ DEBUG = False
 ALLOWED_HOSTS = ["*"]
 # Must set ROOT_URLCONF else we get
 # AttributeError: 'Settings' object has no attribute 'ROOT_URLCONF'
-ROOT_URLCONF = "app.urls"
+ROOT_URLCONF = "djangoapp.urls"
 
 # Turn off deprecation warning
 USE_TZ = True
