+++ _appmap/test/data/django/pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = app.settings
+DJANGO_SETTINGS_MODULE = djangoapp.settings
